+++ Math/Stochastic/Distribution/ExponentialDistribution.php
@@ -107,7 +107,7 @@ class ExponentialDistribution
      */
     public static function getVariance(float $lambda) : float
     {
-        return pow($lambda, -2);
+        return (float) pow($lambda, -2);
     }
 
     /**
